+++ src/main/java/com/teamwizardry/wizardry/api/spell/attribute/AttributeRange.java
@@ -3,12 +3,12 @@ package com.teamwizardry.wizardry.api.spell.attribute;
 public class AttributeRange
 {
 	public static final AttributeRange BACKUP = new AttributeRange(0, 0, 0);
-	
-	public double base;
-	public double min;
-	public double max;
-	
-	public AttributeRange(double base, double min, double max)
+
+	public float base;
+	public float min;
+	public float max;
+
+	public AttributeRange(float base, float min, float max)
 	{
 		this.base = base;
 		this.min = min;
